renamed files and objects to follow the naming conventions.
